# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "description": "A sample product API application with L5 Swagger OpenApi, PHP Pest and Authentication",
        "contact": {
            "email": ""
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api": {},
        "/api/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get list of products",
                "description": "Returns list of products",
                "operationId": "getProductsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "get a single product",
                "description": "",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listing of the products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/products/{id}/variants": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "get a single product with variants",
                "description": "",
                "operationId": "getProductWithLoadingVariantsById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get a single product with variants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductWithVariants"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ProductResource": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductResource"
                }
            },
            "Product": {
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "int64",
                        "default": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Name",
                        "format": "string",
                        "default": "Js Originals relaxed t-shirt with globe back"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "Sku",
                        "format": "string",
                        "default": "204645590"
                    },
                    "barcode": {
                        "title": "barcode",
                        "description": "",
                        "format": "string",
                        "default": "00123456789012"
                    },
                    "options": {
                        "title": "options",
                        "description": "",
                        "format": "json",
                        "default": "['Size']"
                    },
                    "option_values": {
                        "title": "option_values",
                        "description": "",
                        "format": "json",
                        "default": ""
                    },
                    "description": {
                        "title": "description",
                        "description": "Product description",
                        "format": "string",
                        "default": "Js Originals relaxed t-shirt with globe back, 2023 new style!"
                    },
                    "quantity": {
                        "title": "quantity",
                        "description": "quantity",
                        "format": "int",
                        "default": "48"
                    },
                    "price": {
                        "title": "price",
                        "description": "Product price",
                        "format": "float",
                        "default": "13.50"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Product"
                }
            },
            "ProductVariant": {
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "int64",
                        "default": 333
                    },
                    "name": {
                        "title": "name",
                        "description": "Name",
                        "format": "string",
                        "default": "Js Originals relaxed t-shirt with globe back"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "Sku",
                        "format": "string",
                        "default": "204645590"
                    },
                    "barcode": {
                        "title": "barcode",
                        "description": "",
                        "format": "string",
                        "default": "00123456789012"
                    },
                    "options": {
                        "title": "options",
                        "description": "",
                        "format": "json",
                        "default": "['Size']}"
                    },
                    "option_values": {
                        "title": "option_values",
                        "description": "",
                        "format": "json",
                        "default": "['Size' => 'Small']"
                    },
                    "description": {
                        "title": "description",
                        "description": "Product description",
                        "format": "string",
                        "default": "Js Originals relaxed t-shirt with globe back, 2023 new style!"
                    },
                    "quantity": {
                        "title": "quantity",
                        "description": "quantity",
                        "format": "int",
                        "default": "48"
                    },
                    "price": {
                        "title": "price",
                        "description": "Product price",
                        "format": "float",
                        "default": "13.50"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductVariant"
                }
            },
            "ProductWithVariants": {
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "int64",
                        "default": 22
                    },
                    "name": {
                        "title": "name",
                        "description": "Name",
                        "format": "string",
                        "default": "Js Originals relaxed t-shirt with globe back"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "Sku",
                        "format": "string",
                        "default": "204645590"
                    },
                    "barcode": {
                        "title": "barcode",
                        "description": "",
                        "format": "string",
                        "default": "00123456789012"
                    },
                    "options": {
                        "title": "options",
                        "description": "",
                        "format": "json",
                        "default": "['Size']}"
                    },
                    "option_values": {
                        "title": "option_values",
                        "description": "",
                        "format": "json",
                        "default": ""
                    },
                    "description": {
                        "title": "description",
                        "description": "Product description",
                        "format": "string",
                        "default": "Js Originals relaxed t-shirt with globe back, 2023 new style!"
                    },
                    "quantity": {
                        "title": "quantity",
                        "description": "quantity",
                        "format": "int",
                        "default": "23"
                    },
                    "price": {
                        "title": "price",
                        "description": "Product price",
                        "format": "float",
                        "default": "13.50"
                    },
                    "variants": {
                        "title": "variants",
                        "description": "Product variants",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductVariant"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductWithVariants"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "API Endpoints of Projects"
        }
    ]
}